Fix botched merge of PR #246 — main was failing to build
The #246 merge left main with several broken-merge artifacts that
slipped past the failing build-and-release check:

- src/hooks/useGodMode.ts imported adminManager from the deleted
  src/services/subscription.ts (the file was removed by #244).
  Rewritten to read the admin claim via user.getIdTokenResult().
  Behavior matches the pre-#246 inline gate in main.

- src/App.tsx had three sets of duplicate code that the merge
  preserved both copies of:
  1. An inline "1.5/1.6" users-doc + bar-names effect alongside the
     useMyBars() hook call. The inline effect referenced setMyBars /
     setBarDetails / setGlobalNtfyTopic, which no longer exist as
     local setters — fail at tsc.
  2. An orphaned onForegroundMessage(...) block + its cleanup return,
     left over from the pre-extraction push setup, after the
     usePushNotifications() hook call.
  3. An inline admin-claim useEffect AND an inline 60s inactivity
     timer effect with timerRef references, alongside the
     useGodMode() and useInactivityAutoSubmit() hook calls.
  All three duplicates are deleted; hooks own the behavior now.

- src/hooks/useMyBars.ts absorbs the ntfy topic generation that
  used to live inline in App.tsx (#245). Same logic, same
  generateRandomTopicSuffix() helper, same "keep existing topic if
  set" rule. Hook now returns { myBars, barDetails, globalNtfyTopic }.

- src/hooks/useAuth.ts: drop an unused useCallback import.

After this: tsc --noEmit clean, 48/48 jsdom tests pass, vite build
succeeds.
