+++ src/hooks/useAuth.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 import {
   onAuthStateChanged,
   signInWithPopup,
