+++ src/hooks/useGodMode.ts
@@ -1,24 +1,31 @@
 import { useEffect, useState } from 'react';
 import type { User } from 'firebase/auth';
-import { adminManager } from '../services/subscription';
 
 // Resolves whether the current user has elevated ("god mode")
-// privileges. NOTE: this currently routes through `adminManager`,
-// which checks `user.email` against VITE_GOD_MODE_EMAIL — a
-// client-side gate that ships the admin address in the bundle. The
-// rules in firestore.rules do NOT enforce this, so anything gated by
-// the returned value should be treated as UI affordance only. A
-// follow-up should replace this with a Firebase Auth custom claim
-// read via getIdTokenResult().
+// privileges by reading the `admin` custom claim off their ID token.
+// Claims are signed by Firebase so they cannot be spoofed from the
+// client, and the same claim is checked in firestore.rules
+// (request.auth.token.admin). Set the claim with
+// `node scripts/set-admin-claim.js --email ...`.
 export function useGodMode(user: User | null) {
   const [isGodMode, setIsGodMode] = useState(false);
 
   useEffect(() => {
+    let cancelled = false;
     if (!user) {
       setIsGodMode(false);
       return;
     }
-    setIsGodMode(adminManager.checkSubscription(user.email));
+    user.getIdTokenResult()
+      .then(result => {
+        if (cancelled) return;
+        setIsGodMode(result.claims.admin === true);
+      })
+      .catch(err => {
+        console.warn('Failed to read admin claim', err);
+        if (!cancelled) setIsGodMode(false);
+      });
+    return () => { cancelled = true; };
   }, [user]);
 
   return isGodMode;
